requirements.txt
mindsdb_sdk
python-dotenv
pandas
Flask